mptcp: don't reset sk_err after sk_error_report
Currently MPTCP reset the value of sk_err to 0 after the callback to
sk_error_report. This is not needed and not done like that in TCP
stack.

If we don't reset sk_err, we will still be able to get this info when
deleting the subflow (in mptcp_del_sock).

Signed-off-by: Gregory Detal <gregory.detal@tessares.net>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>